requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
PyMuPDF==1.23.0
celery==5.3.4
redis==5.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-Uploads==0.2.1
PyPDF2==3.0.1
reportlab==4.0.4
pdf2docx==0.5.6
pdf2image==1.16.3
tabula-py==2.7.0
python-pptx==0.6.21
openpyxl==3.1.2
Pillow==10.0.0